refactor: split expression inference
