+++ src/types/checker/inference/expr/assignments.rs
@@ -1,0 +1,68 @@
+use crate::errors::CompileError;
+use crate::parser::ast::{Expr, ExprKind, Stmt, StmtKind};
+use crate::span::Span;
+use crate::types::{PhpType, TypeEnv};
+
+use super::super::super::Checker;
+
+impl Checker {
+    pub(super) fn check_assignment_expression(
+        &mut self,
+        target: &Expr,
+        value: &Expr,
+        result_target: Option<&Expr>,
+        prelude: &[Stmt],
+        span: Span,
+        env: &mut TypeEnv,
+    ) -> Result<PhpType, CompileError> {
+        for stmt in prelude {
+            self.check_assignment_like_stmt(stmt, env)?;
+        }
+
+        if let ExprKind::Variable(name) = &target.kind {
+            return self.check_local_assignment_expression(name, value, span, env);
+        }
+
+        let stmt_kind = match &target.kind {
+            ExprKind::ArrayAccess { array, index } => match &array.kind {
+                ExprKind::Variable(array) => StmtKind::ArrayAssign {
+                    array: array.clone(),
+                    index: *index.clone(),
+                    value: value.clone(),
+                },
+                ExprKind::PropertyAccess { object, property } => StmtKind::PropertyArrayAssign {
+                    object: object.clone(),
+                    property: property.clone(),
+                    index: *index.clone(),
+                    value: value.clone(),
+                },
+                ExprKind::StaticPropertyAccess { receiver, property } => {
+                    StmtKind::StaticPropertyArrayAssign {
+                        receiver: receiver.clone(),
+                        property: property.clone(),
+                        index: *index.clone(),
+                        value: value.clone(),
+                    }
+                }
+                _ => return Err(CompileError::new(span, "Invalid assignment target")),
+            },
+            ExprKind::PropertyAccess { object, property } => StmtKind::PropertyAssign {
+                object: object.clone(),
+                property: property.clone(),
+                value: value.clone(),
+            },
+            ExprKind::StaticPropertyAccess { receiver, property } => {
+                StmtKind::StaticPropertyAssign {
+                    receiver: receiver.clone(),
+                    property: property.clone(),
+                    value: value.clone(),
+                }
+            }
+            _ => return Err(CompileError::new(span, "Invalid assignment target")),
+        };
+
+        let stmt = Stmt::new(stmt_kind, span);
+        self.check_assignment_like_stmt(&stmt, env)?;
+        self.infer_type(result_target.unwrap_or(target), env)
+    }
+}
