+++ src/types/checker/inference/expr/class_refs.rs
@@ -1,0 +1,85 @@
+use crate::errors::CompileError;
+use crate::parser::ast::{Expr, StaticReceiver};
+use crate::span::Span;
+use crate::types::TypeEnv;
+
+use super::super::super::Checker;
+
+impl Checker {
+    pub(super) fn validate_late_bound_constructor_targets(
+        &mut self,
+        base_class: &str,
+        args: &[Expr],
+        expr: &Expr,
+        env: &TypeEnv,
+    ) -> Result<(), CompileError> {
+        let mut class_names: Vec<String> = self
+            .classes
+            .keys()
+            .filter(|name| self.class_is_same_or_descends_from(name, base_class))
+            .cloned()
+            .collect();
+        class_names.sort();
+
+        for class_name in class_names {
+            self.infer_new_object_type(&class_name, args, expr, env)?;
+        }
+
+        Ok(())
+    }
+
+    fn class_is_same_or_descends_from(&self, class_name: &str, base_class: &str) -> bool {
+        let mut current = Some(class_name);
+        while let Some(name) = current {
+            if name == base_class {
+                return true;
+            }
+            current = self
+                .classes
+                .get(name)
+                .and_then(|info| info.parent.as_deref());
+        }
+        false
+    }
+
+    pub(super) fn validate_class_constant_receiver(
+        &self,
+        receiver: &StaticReceiver,
+        span: Span,
+    ) -> Result<(), CompileError> {
+        match receiver {
+            StaticReceiver::Named(_) => Ok(()),
+            StaticReceiver::Self_ | StaticReceiver::Static => {
+                if self.current_class.is_some() {
+                    Ok(())
+                } else {
+                    Err(CompileError::new(
+                        span,
+                        "Cannot use self::class or static::class outside a class context",
+                    ))
+                }
+            }
+            StaticReceiver::Parent => {
+                let current = self.current_class.as_ref().ok_or_else(|| {
+                    CompileError::new(
+                        span,
+                        "Cannot use parent::class outside a class context",
+                    )
+                })?;
+                if self
+                    .classes
+                    .get(current)
+                    .and_then(|info| info.parent.as_ref())
+                    .is_some()
+                {
+                    Ok(())
+                } else {
+                    Err(CompileError::new(
+                        span,
+                        &format!("Class '{}' has no parent class", current),
+                    ))
+                }
+            }
+        }
+    }
+}
